recommendatinos cron finished

+++ internal/service/activity/model.go
@@ -8,6 +8,6 @@ type Activity struct {
 	CreatedAt *time.Time `db:"created_at"`
 	UpdatedAt *time.Time `db:"updated_at"`
 	Address   string     `db:"address"`
-	Lon       float32    `db:"lon"`
-	Lat       float32    `db:"lat"`
+	Lon       float64    `db:"lon"`
+	Lat       float64    `db:"lat"`
 }
