+++ internal/service/activity_interest/model.go
@@ -5,7 +5,7 @@ import "time"
 type ActivityInterest struct {
 	ActivityId int64      `db:"activity_id"`
 	InterestId int64      `db:"interest_id"`
-	Rank       float32    `db:"rank"`
+	Rank       float64    `db:"rank"`
 	CreatedAt  *time.Time `db:"created_at"`
 	UpdatedAt  *time.Time `db:"updated_at"`
 }
